SZEPULHU-6: Implement client registration
The client can register using the top navigation menu.
When the registration has been completed, the client get to the confirmation page and be able to use the website as logged in user.
Add first naem and last name fields to all type of users.
Create a new Client context for behat tests and related page objects.
Enable Xdebug when using the "behat" command.