SZEPULHU-9: Write Behat test for view featured professionals scenario
Write the ElementFinderExtension which helps to build a filter conditions to find complex element on the page.
Use this extension together with PageObjectExtension.
Create a new fixture "professional_featured" to have something to check on the homepage.
Fix displaying name of featured professionals according to the current language (order fo first, last name).