SZEPULHU-9: Do not store location data in cookie
The featured professionals on the homepage is listed based on the location given by the HTML5 geolocation API. Current implementation store this value in a cookie and do not call the geolocation API any more after the first call. Cookie is also updated when the user changes the location (city or county).

With this change the geolocation API will be called on every page load. If the user changes the location by selecting a different city or county, it will be reset at next page load.
When geolocation API is not available or Javascript is disabled in the browser, the capital will be set as default location.

Changes:
- move location element into the action "embeddedFeaturedProfessionalsAction"
- as result of the previous change move featued professional specific code from HomepageResponse to FeaturedProfessionalResponse
- add autocomplete="off" attribute on the location select, because of a Firefox bug, see http://stackoverflow.com/questions/4831848/firefox-ignores-option-selected-selected
- refactor Javascript code used for loading featured professionals