SZEPULHU-10: Let Xdebug step into page objects
Page object extension uses a proxy library which will generate proxy classes on the fly for each page object class. These are saved into the tmp directory on the server, making Xdebug unable to access it's source for debugging.
This commit shares the tmp with the host, so a path mapping can be set to make Xdebug working.